split out hash table store from dht, wip #434
q
