+++ dht.go
@@ -1,4 +1,4 @@
-// Copyright (C) 2013-2017, The MetaCurrency Project (Eric Harris-Braun, Arthur Brock, et. al.)
+// Copyright (C) 2013-2018, The MetaCurrency Project (Eric Harris-Braun, Arthur Brock, et. al.)
 // Use of this source code is governed by GPLv3 found in the LICENSE file
 //----------------------------------------------------------------------------------------
 
@@ -7,19 +7,13 @@
 package holochain
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"path/filepath"
-	"strconv"
-	"strings"
 	"sync"
 
 	. "github.com/holochain/holochain-proto/hash"
 	peer "github.com/libp2p/go-libp2p-peer"
-	"github.com/tidwall/buntdb"
 )
 
 // Holds the dht configuration options
@@ -54,7 +48,7 @@ type Channel chan interface{}
 // DHT struct holds the data necessary to run the distributed hash table
 type DHT struct {
 	h          *Holochain // pointer to the holochain this DHT is part of
-	db         *buntdb.DB
+	ht         HashTable
 	retryQueue chan *retry
 	gossipPuts Channel
 	glog       *Logger // the gossip logger
@@ -73,58 +67,6 @@ type retry struct {
 
 const (
 	MaxRetries = 10
-)
-
-// Meta holds data that can be associated with a hash
-// @todo, we should also be storing the meta-data source
-type Meta struct {
-	H Hash   // hash of link-data associated
-	T string // meta-data type identifier
-	V []byte // meta-data
-}
-
-const (
-	// constants for status action type
-
-	AddLinkAction = ""
-	DelLinkAction = "d"
-
-	// constants for the state of the data, they are bit flags
-
-	StatusDefault  = 0x00
-	StatusLive     = 0x01
-	StatusRejected = 0x02
-	StatusDeleted  = 0x04
-	StatusModified = 0x08
-	StatusAny      = 0xFF
-
-	// constants for the stored string status values in buntdb and for building code
-
-	StatusLiveVal     = "1"
-	StatusRejectedVal = "2"
-	StatusDeletedVal  = "4"
-	StatusModifiedVal = "8"
-	StatusAnyVal      = "255"
-
-	// constants for system reseved tags (start with 2 underscores)
-
-	SysTagReplacedBy = "__replacedBy"
-
-	// constants for get request GetMask
-
-	GetMaskDefault   = 0x00
-	GetMaskEntry     = 0x01
-	GetMaskEntryType = 0x02
-	GetMaskSources   = 0x04
-	GetMaskAll       = 0xFF
-
-	// constants for building code for GetMask
-
-	GetMaskDefaultStr   = "0"
-	GetMaskEntryStr     = "1"
-	GetMaskEntryTypeStr = "2"
-	GetMaskSourcesStr   = "4"
-	GetMaskAllStr       = "255"
 )
 
 // HoldReq holds the data of a put request
@@ -193,15 +135,6 @@ type GetLinksOptions struct {
 	StatusMask int  // mask of which status of links to return
 }
 
-// TaggedHash holds associated entries for the LinkQueryResponse
-type TaggedHash struct {
-	H         string // the hash of the link; gets filled by dht base node when answering get link request
-	E         string // the value of link, gets filled if options set Load to true
-	EntryType string // the entry type of the link, gets filled if options set Load to true
-	T         string // the tag of the link, gets filled only if a tag wasn't specified and all tags are being returns
-	Source    string // the statuses on the link, gets filled if options set Load to true
-}
-
 // LinkQueryResp holds response to getLinks query
 type LinkQueryResp struct {
 	Links []TaggedHash
@@ -213,22 +146,6 @@ type ListAddReq struct {
 	WarrantType int
 	Warrant     []byte
 }
-
-// LinkEvent represents the value stored in buntDB associated with a
-// link key for one source having stored one LinkingEntry
-// (The Link struct defined in entry.go is encoded in the key used for buntDB)
-type LinkEvent struct {
-	Status     int
-	Source     string
-	LinksEntry string
-}
-
-var ErrLinkNotFound = errors.New("link not found")
-var ErrPutLinkOverDeleted = errors.New("putlink over deleted link")
-var ErrHashDeleted = errors.New("hash deleted")
-var ErrHashModified = errors.New("hash modified")
-var ErrHashRejected = errors.New("hash rejected")
-var ErrEntryTypeMismatch = errors.New("entry type mismatch")
 
 var KValue int = 10
 var AlphaValue int = 3
@@ -246,17 +163,8 @@ func NewDHT(h *Holochain) *DHT {
 		dlog:   &h.Config.Loggers.DHT,
 		config: &h.Nucleus().DNA().DHTConfig,
 	}
-	db, err := buntdb.Open(filepath.Join(h.DBPath(), DHTStoreFileName))
-	if err != nil {
-		panic(err)
-	}
-	db.CreateIndex("link", "link:*", buntdb.IndexString)
-	db.CreateIndex("idx", "idx:*", buntdb.IndexInt)
-	db.CreateIndex("peer", "peer:*", buntdb.IndexString)
-	db.CreateIndex("list", "list:*", buntdb.IndexString)
-	db.CreateIndex("entry", "entry:*", buntdb.IndexString)
-
-	dht.db = db
+	dht.ht = &BuntHT{}
+	dht.ht.Open(filepath.Join(h.DBPath(), DHTStoreFileName))
 	dht.retryQueue = make(chan *retry, 100)
 
 	//	dht.sources = make(map[peer.ID]bool)
@@ -286,7 +194,7 @@ func (dht *DHT) putKey(agent Agent) (err error) {
 	if err != nil {
 		return
 	}
-	if err = dht.put(dht.h.node.NewMessage(PUT_REQUEST, PutReq{H: keyHash}), KeyEntryType, keyHash, nodeID, []byte(pubKey), StatusLive); err != nil {
+	if err = dht.Put(dht.h.node.NewMessage(PUT_REQUEST, PutReq{H: keyHash}), KeyEntryType, keyHash, nodeID, []byte(pubKey), StatusLive); err != nil {
 		return
 	}
 	return
@@ -297,7 +205,7 @@ func (dht *DHT) SetupDHT() (err error) {
 	x := ""
 	// put the holochain id so it always exists for linking
 	dna := dht.h.DNAHash()
-	err = dht.put(nil, DNAEntryType, dna, dht.h.nodeID, []byte(x), StatusLive)
+	err = dht.Put(nil, DNAEntryType, dna, dht.h.nodeID, []byte(x), StatusLive)
 	if err != nil {
 		return
 	}
@@ -323,354 +231,76 @@ func (dht *DHT) SetupDHT() (err error) {
 	if err != nil {
 		return
 	}
-	if err = dht.put(dht.h.node.NewMessage(PUT_REQUEST, PutReq{H: a}), AgentEntryType, a, dht.h.nodeID, b, StatusLive); err != nil {
+	if err = dht.Put(dht.h.node.NewMessage(PUT_REQUEST, PutReq{H: a}), AgentEntryType, a, dht.h.nodeID, b, StatusLive); err != nil {
 		return
 	}
 
 	return
 }
 
-// put stores a value to the DHT store
+// Put stores a value to the DHT store
 // N.B. This call assumes that the value has already been validated
-func (dht *DHT) put(m *Message, entryType string, key Hash, src peer.ID, value []byte, status int) (err error) {
-	k := key.String()
-	dht.dlog.Logf("put %s=>%s", k, string(value))
-	err = dht.db.Update(func(tx *buntdb.Tx) error {
-		_, err := incIdx(tx, m)
-		if err != nil {
-			return err
-		}
-		_, _, err = tx.Set("entry:"+k, string(value), nil)
-		if err != nil {
-			return err
-		}
-		_, _, err = tx.Set("type:"+k, entryType, nil)
-		if err != nil {
-			return err
-		}
-		_, _, err = tx.Set("src:"+k, peer.IDB58Encode(src), nil)
-		if err != nil {
-			return err
-		}
-		_, _, err = tx.Set("status:"+k, fmt.Sprintf("%d", status), nil)
-		if err != nil {
-			return err
-		}
-		return err
-	})
+func (dht *DHT) Put(m *Message, entryType string, key Hash, src peer.ID, value []byte, status int) (err error) {
+	dht.dlog.Logf("put %v=>%s", key, string(value))
+	err = dht.ht.Put(m, entryType, key, src, value, status)
 	return
 }
 
-func _setStatus(tx *buntdb.Tx, m *Message, key string, status int) (err error) {
-
-	_, err = tx.Get("entry:" + key)
-	if err != nil {
-		if err == buntdb.ErrNotFound {
-			err = ErrHashNotFound
-		}
-		return
-	}
-
-	_, err = incIdx(tx, m)
-	if err != nil {
-		return
-	}
-
-	_, _, err = tx.Set("status:"+key, fmt.Sprintf("%d", status), nil)
-	if err != nil {
-		return
-	}
-	return
-}
-
-// del moves the given hash to the StatusDeleted status
+// Del moves the given hash to the StatusDeleted status
 // N.B. this functions assumes that the validity of this action has been confirmed
-func (dht *DHT) del(m *Message, key Hash) (err error) {
-	k := key.String()
-	dht.dlog.Logf("del %s", k)
-	err = dht.db.Update(func(tx *buntdb.Tx) error {
-		err = _setStatus(tx, m, k, StatusDeleted)
-		return err
-	})
+func (dht *DHT) Del(m *Message, key Hash) (err error) {
+	dht.dlog.Logf("del %v", key)
+	err = dht.ht.Del(m, key)
 	return
 }
 
-// mod moves the given hash to the StatusModified status
+// Mod moves the given hash to the StatusModified status
 // N.B. this functions assumes that the validity of this action has been confirmed
-func (dht *DHT) mod(m *Message, key Hash, newkey Hash) (err error) {
-	k := key.String()
-	dht.dlog.Logf("mod %s", k)
-	err = dht.db.Update(func(tx *buntdb.Tx) error {
-		err = _setStatus(tx, m, k, StatusModified)
-		if err == nil {
-			link := newkey.String()
-			err = _link(tx, k, link, SysTagReplacedBy, m.From, StatusLive, newkey)
-			if err == nil {
-				_, _, err = tx.Set("replacedBy:"+k, link, nil)
-				if err != nil {
-					return err
-				}
-			}
-		}
-		return err
-	})
+func (dht *DHT) Mod(m *Message, key Hash, newkey Hash) (err error) {
+	dht.dlog.Logf("mod %v", key)
+	err = dht.ht.Mod(m, key, newkey)
 	return
 }
 
-func _get(tx *buntdb.Tx, k string, statusMask int) (string, error) {
-	val, err := tx.Get("entry:" + k)
-	if err == buntdb.ErrNotFound {
-		err = ErrHashNotFound
-		return val, err
-	}
-	var statusVal string
-	statusVal, err = tx.Get("status:" + k)
-	if err == nil {
-
-		if statusMask == StatusDefault {
-			// if the status mask is not given (i.e. Default) then
-			// we return information about the status if it's other than live
-			switch statusVal {
-			case StatusDeletedVal:
-				err = ErrHashDeleted
-			case StatusModifiedVal:
-				val, err = tx.Get("replacedBy:" + k)
-				if err != nil {
-					panic("missing expected replacedBy record")
-				}
-				err = ErrHashModified
-			case StatusRejectedVal:
-				err = ErrHashRejected
-			case StatusLiveVal:
-			default:
-				panic("unknown status!")
-			}
-		} else {
-			// otherwise we return the value only if the status is in the mask
-			var status int
-			status, err = strconv.Atoi(statusVal)
-			if err == nil {
-				if (status & statusMask) == 0 {
-					err = ErrHashNotFound
-				}
-			}
-		}
-	}
-	return val, err
-}
-
-// exists checks for the existence of the hash in the store
-func (dht *DHT) exists(key Hash, statusMask int) (err error) {
-	err = dht.db.View(func(tx *buntdb.Tx) error {
-		_, err := _get(tx, key.String(), statusMask)
-		return err
-	})
+// Exists checks for the existence of the hash in the store
+func (dht *DHT) Exists(key Hash, statusMask int) (err error) {
+	err = dht.ht.Exists(key, statusMask)
 	return
 }
 
-// returns the source of a given hash
-func (dht *DHT) source(key Hash) (id peer.ID, err error) {
-	err = dht.db.View(func(tx *buntdb.Tx) error {
-		val, err := tx.Get("src:" + key.String())
-		if err == buntdb.ErrNotFound {
-			err = ErrHashNotFound
-		}
-		if err == nil {
-			id, err = peer.IDB58Decode(val)
-		}
-		return err
-	})
+// Source returns the source node address of a given hash
+func (dht *DHT) Source(key Hash) (id peer.ID, err error) {
+	id, err = dht.ht.Source(key)
 	return
 }
 
-// get retrieves a value from the DHT store
-func (dht *DHT) get(key Hash, statusMask int, getMask int) (data []byte, entryType string, sources []string, status int, err error) {
-	if getMask == GetMaskDefault {
-		getMask = GetMaskEntry
-	}
-	err = dht.db.View(func(tx *buntdb.Tx) error {
-		k := key.String()
-		val, err := _get(tx, k, statusMask)
-		if err != nil {
-			data = []byte(val) // gotta do this because value is valid if ErrHashModified
-			return err
-		}
-		data = []byte(val)
-
-		if (getMask & GetMaskEntryType) != 0 {
-			entryType, err = tx.Get("type:" + k)
-			if err != nil {
-				return err
-			}
-		}
-		if (getMask & GetMaskSources) != 0 {
-			val, err = tx.Get("src:" + k)
-			if err == buntdb.ErrNotFound {
-				err = ErrHashNotFound
-			}
-			if err == nil {
-				sources = append(sources, val)
-			}
-			if err != nil {
-				return err
-			}
-		}
-
-		val, err = tx.Get("status:" + k)
-		if err != nil {
-			return err
-		}
-		status, err = strconv.Atoi(val)
-		if err != nil {
-			return err
-		}
-
-		return err
-	})
+// Get retrieves a value from the DHT store
+func (dht *DHT) Get(key Hash, statusMask int, getMask int) (data []byte, entryType string, sources []string, status int, err error) {
+	data, entryType, sources, status, err = dht.ht.Get(key, statusMask, getMask)
 	return
 }
 
-// _link is a low level routine to add a link, also used by delLink
-// this ensure monotonic recording of linking attempts
-func _link(tx *buntdb.Tx, base string, link string, tag string, src peer.ID, status int, linkingEntryHash Hash) (err error) {
-	key := "link:" + base + ":" + link + ":" + tag
-	var val string
-	val, err = tx.Get(key)
-	source := peer.IDB58Encode(src)
-	lehStr := linkingEntryHash.String()
-	var records []LinkEvent
-	if err == nil {
-		// load the previous value so we can append to it.
-		json.Unmarshal([]byte(val), &records)
-
-		// TODO: if the link exists, then load the statuses and see
-		// what we should do about this situation
-		/*
-			// search for the source and linking entry in the status
-			for _, s := range records {
-				if s.Source == source && s.LinksEntry == lehStr {
-					if status == StatusLive && s.Status != status {
-						err = ErrPutLinkOverDeleted
-						return
-					}
-					// return silently because this is just a duplicate putLink
-					break
-				}
-			} // fall through and add this linking event.
-		*/
-
-	} else if err == buntdb.ErrNotFound {
-		// when deleting the key must exist
-		if status == StatusDeleted {
-			err = ErrLinkNotFound
-			return
-		}
-		err = nil
-	} else {
-		return
-	}
-	records = append(records, LinkEvent{status, source, lehStr})
-	var b []byte
-	b, err = json.Marshal(records)
-	if err != nil {
-		return
-	}
-	_, _, err = tx.Set(key, string(b), nil)
-	if err != nil {
-		return
-	}
-	return
-}
-
-func (dht *DHT) link(m *Message, base string, link string, tag string, status int) (err error) {
-	err = dht.db.Update(func(tx *buntdb.Tx) error {
-		_, err := _get(tx, base, StatusLive)
-		if err != nil {
-			return err
-		}
-		err = _link(tx, base, link, tag, m.From, status, m.Body.(LinkReq).Links)
-		if err != nil {
-			return err
-		}
-
-		//var index string
-		_, err = incIdx(tx, m)
-
-		if err != nil {
-			return err
-		}
-		return nil
-	})
-	return
-}
-
-// putLink associates a link with a stored hash
+// PutLink associates a link with a stored hash
 // N.B. this function assumes that the data associated has been properly retrieved
 // and validated from the cource chain
-func (dht *DHT) putLink(m *Message, base string, link string, tag string) (err error) {
+func (dht *DHT) PutLink(m *Message, base string, link string, tag string) (err error) {
 	dht.dlog.Logf("putLink on %v link %v as %s", base, link, tag)
-	err = dht.link(m, base, link, tag, StatusLive)
+	err = dht.ht.PutLink(m, base, link, tag)
 	return
 }
 
-// delLink removes a link and tag associated with a stored hash
+// DelLink removes a link and tag associated with a stored hash
 // N.B. this function assumes that the action has been properly validated
-func (dht *DHT) delLink(m *Message, base string, link string, tag string) (err error) {
+func (dht *DHT) DelLink(m *Message, base string, link string, tag string) (err error) {
 	dht.dlog.Logf("delLink on %v link %v as %s", base, link, tag)
-	err = dht.link(m, base, link, tag, StatusDeleted)
+	err = dht.ht.DelLink(m, base, link, tag)
 	return
 }
 
-func filter(ss []Meta, test func(*Meta) bool) (ret []Meta) {
-	for _, s := range ss {
-		if test(&s) {
-			ret = append(ret, s)
-		}
-	}
-	return
-}
-
-// getLinks retrieves meta value associated with a base
-func (dht *DHT) getLinks(base Hash, tag string, statusMask int) (results []TaggedHash, err error) {
+// GetLinks retrieves meta value associated with a base
+func (dht *DHT) GetLinks(base Hash, tag string, statusMask int) (results []TaggedHash, err error) {
 	dht.dlog.Logf("getLinks on %v of %s with mask %d", base, tag, statusMask)
-	b := base.String()
-	err = dht.db.View(func(tx *buntdb.Tx) error {
-		_, err := _get(tx, b, StatusLive+StatusModified) //only get links on live and modified bases
-		if err != nil {
-			return err
-		}
-
-		if statusMask == StatusDefault {
-			statusMask = StatusLive
-		}
-
-		results = make([]TaggedHash, 0)
-		err = tx.Ascend("link", func(key, value string) bool {
-			x := strings.Split(key, ":")
-			t := string(x[3])
-			if string(x[1]) == b && (tag == "" || tag == t) {
-				var records []LinkEvent
-				json.Unmarshal([]byte(value), &records)
-				l := len(records)
-				//TODO: this is totally bogus currently simply
-				// looking at the last item we ever got
-				if l > 0 {
-					entry := records[l-1]
-					if err == nil && (entry.Status&statusMask) > 0 {
-						th := TaggedHash{H: string(x[2]), Source: entry.Source}
-						if tag == "" {
-							th.T = t
-						}
-						results = append(results, th)
-					}
-				}
-			}
-
-			return true
-		})
-
-		return err
-	})
+	results, err = dht.ht.GetLinks(base, tag, statusMask)
 	return
 }
 
@@ -815,189 +445,15 @@ func (dht *DHT) Start() (err error) {
 	return
 }
 
-// DumpIdx converts message and data of a DHT change request to a string for human consumption
-func (dht *DHT) DumpIdx(idx int) (str string, err error) {
-	var msg Message
-	msg, err = dht.GetIdxMessage(idx)
-	if err != nil {
-		return
-	}
-	f, _ := msg.Fingerprint()
-	str = fmt.Sprintf("MSG (fingerprint %v):\n   %v\n", f, msg)
-	switch msg.Type {
-	case PUT_REQUEST:
-		key := msg.Body.(PutReq).H
-		entry, entryType, _, _, e := dht.get(key, StatusDefault, GetMaskAll)
-		if e != nil {
-			err = fmt.Errorf("couldn't get %v err:%v ", key, e)
-			return
-		} else {
-			str += fmt.Sprintf("DATA: type:%s entry: %v\n", entryType, entry)
-		}
-	}
-	return
-}
-
-func statusValueToString(val string) string {
-	//TODO
-	return val
-}
-
 // String converts a DHT into a human readable string
 func (dht *DHT) String() (result string) {
-	idx, err := dht.GetIdx()
-	if err != nil {
-		return err.Error()
-	}
-	result += fmt.Sprintf("DHT changes: %d\n", idx)
-	for i := 1; i <= idx; i++ {
-		str, err := dht.DumpIdx(i)
-		if err != nil {
-			result += fmt.Sprintf("%d Error:%v\n", i, err)
-		} else {
-			result += fmt.Sprintf("%d\n%v\n", i, str)
-		}
-	}
-
-	result += fmt.Sprintf("DHT entries:\n")
-	err = dht.db.View(func(tx *buntdb.Tx) error {
-		err = tx.Ascend("entry", func(key, value string) bool {
-			x := strings.Split(key, ":")
-			k := string(x[1])
-			var status string
-			statusVal, err := tx.Get("status:" + k)
-			if err != nil {
-				status = fmt.Sprintf("<err getting status:%v>", err)
-			} else {
-				status = statusValueToString(statusVal)
-			}
-
-			var sources string
-			sources, err = tx.Get("src:" + k)
-			if err != nil {
-				sources = fmt.Sprintf("<err getting sources:%v>", err)
-			}
-			var links string
-			err = tx.Ascend("link", func(key, value string) bool {
-				x := strings.Split(key, ":")
-				base := x[1]
-				link := x[2]
-				tag := x[3]
-				if base == k {
-					links += fmt.Sprintf("Linked to: %s with tag %s\n", link, tag)
-					links += value + "\n"
-				}
-				return true
-			})
-			result += fmt.Sprintf("Hash--%s (status %s):\nValue: %s\nSources: %s\n%s\n", k, status, value, sources, links)
-			return true
-		})
-		return nil
-	})
-
-	return
-}
-
-// DumpIdxJSON converts message and data of a DHT change request to a JSON string representation.
-func (dht *DHT) DumpIdxJSON(idx int) (str string, err error) {
-	var msg Message
-	var buffer bytes.Buffer
-	var msgField, dataField string
-	msg, err = dht.GetIdxMessage(idx)
-
-	if err != nil {
-		return "", err
-	}
-
-	f, _ := msg.Fingerprint()
-	buffer.WriteString(fmt.Sprintf("{ \"index\": %d,", idx))
-	msgField = fmt.Sprintf("\"message\": { \"fingerprint\": \"%v\", \"content\": \"%v\" },", f, msg)
-
-	switch msg.Type {
-	case PUT_REQUEST:
-		key := msg.Body.(PutReq).H
-		entry, entryType, _, _, e := dht.get(key, StatusDefault, GetMaskAll)
-		if e != nil {
-			err = fmt.Errorf("couldn't get %v err:%v ", key, e)
-			return
-		}
-		dataField = fmt.Sprintf("\"data\": { \"type\": \"%s\", \"entry\": \"%v\" }", entryType, entry)
-	}
-
-	if len(dataField) > 0 {
-		buffer.WriteString(msgField)
-		buffer.WriteString(dataField)
-	} else {
-		buffer.WriteString(strings.TrimSuffix(msgField, ","))
-	}
-	buffer.WriteString("}")
-	return PrettyPrintJSON(buffer.Bytes())
+	return dht.ht.String()
 }
 
 // JSON converts a DHT into a JSON string representation.
 func (dht *DHT) JSON() (result string, err error) {
-	var buffer, entries bytes.Buffer
-	idx, err := dht.GetIdx()
-	if err != nil {
-		return "", err
-	}
-	buffer.WriteString("{ \"dht_changes\": [")
-	for i := 1; i <= idx; i++ {
-		json, err := dht.DumpIdxJSON(i)
-		if err != nil {
-			return "", fmt.Errorf("DHT Change %d,  Error: %v", i, err)
-		}
-		buffer.WriteString(json)
-		if i < idx {
-			buffer.WriteString(",")
-		}
-	}
-	buffer.WriteString("], \"dht_entries\": [")
-	err = dht.db.View(func(tx *buntdb.Tx) error {
-		err = tx.Ascend("entry", func(key, value string) bool {
-			x := strings.Split(key, ":")
-			k := string(x[1])
-			var status string
-			statusVal, err := tx.Get("status:" + k)
-			if err != nil {
-				status = fmt.Sprintf("<err getting status:%v>", err)
-			} else {
-				status = statusValueToString(statusVal)
-			}
-
-			var sources string
-			sources, err = tx.Get("src:" + k)
-			if err != nil {
-				sources = fmt.Sprintf("<err getting sources:%v>", err)
-			}
-			var links bytes.Buffer
-			err = tx.Ascend("link", func(key, value string) bool {
-				x := strings.Split(key, ":")
-				base := x[1]
-				link := x[2]
-				tag := x[3]
-				if base == k {
-					links.WriteString(fmt.Sprintf("{ \"linkTo\": \"%s\",", link))
-					links.WriteString(fmt.Sprintf("\"tag\": \"%s\",", tag))
-					links.WriteString(fmt.Sprintf("\"value\": \"%s\" },", EscapeJSONValue(value)))
-				}
-				return true
-			})
-			entries.WriteString(fmt.Sprintf("{ \"hash\": \"%s\",", k))
-			entries.WriteString(fmt.Sprintf("\"status\": \"%s\",", status))
-			entries.WriteString(fmt.Sprintf("\"value\": \"%s\",", EscapeJSONValue(value)))
-			entries.WriteString(fmt.Sprintf("\"sources\": \"%s\"", sources))
-			if links.Len() > 0 {
-				entries.WriteString(fmt.Sprintf(",\"links\": [%s]", strings.TrimSuffix(links.String(), ",")))
-			}
-			entries.WriteString("},")
-			return true
-		})
-		return nil
-	})
-	buffer.WriteString(strings.TrimSuffix(entries.String(), ","))
-	buffer.WriteString("]}")
-	return PrettyPrintJSON(buffer.Bytes())
+	result, err = dht.ht.JSON()
+	return
 }
 
 // Close cleans up the DHT
@@ -1008,8 +464,19 @@ func (dht *DHT) Close() {
 	dht.gchan = nil
 	close(dht.gossipPuts)
 	dht.gossipPuts = nil
-	dht.db.Close()
-	dht.db = nil
+	dht.ht.Close()
+}
+
+// GetIdx returns the current index of changes to the HashTable
+func (dht *DHT) GetIdx() (idx int, err error) {
+	idx, err = dht.ht.GetIdx()
+	return
+}
+
+// GetIdxMessage returns the messages that causes the change at a given index
+func (dht *DHT) GetIdxMessage(idx int) (msg Message, err error) {
+	msg, err = dht.ht.GetIdxMessage(idx)
+	return
 }
 
 // Retry starts retry processing
