+++ holochain_test.go
@@ -501,7 +501,7 @@ func TestCommit(t *testing.T) {
 	hash := commit(h, "oddNumbers", "7")
 
 	Convey("publicly shared entries should generate a put", t, func() {
-		err := h.dht.exists(hash, StatusLive)
+		err := h.dht.Exists(hash, StatusLive)
 		So(err, ShouldBeNil)
 	})
 
@@ -510,7 +510,7 @@ func TestCommit(t *testing.T) {
 	Convey("it should attach links after commit of Links entry", t, func() {
 		commit(h, "rating", fmt.Sprintf(`{"Links":[{"Base":"%s","Link":"%s","Tag":"4stars"}]}`, hash.String(), profileHash.String()))
 
-		results, err := h.dht.getLinks(hash, "4stars", StatusLive)
+		results, err := h.dht.GetLinks(hash, "4stars", StatusLive)
 		So(err, ShouldBeNil)
 		So(fmt.Sprintf("%v", results), ShouldEqual, fmt.Sprintf("[{QmYeinX5vhuA91D3v24YbgyLofw9QAxY6PoATrBHnRwbtt    %s}]", h.nodeIDStr))
 	})
