+++ ht_test.go
@@ -1,0 +1,12 @@
+package holochain
+
+import (
+	//	. "github.com/holochain/holochain-proto/hash"
+	. "github.com/smartystreets/goconvey/convey"
+	"testing"
+)
+
+func Test(t *testing.T) {
+	Convey("", t, func() {
+	})
+}
