+++ jsribosome_test.go
@@ -897,9 +897,9 @@ func TestJSDHT(t *testing.T) {
 		_, err = NewHash(z.lastResult.String())
 		So(err, ShouldBeNil)
 
-		links, _ := h.dht.getLinks(hash, "4stars", StatusLive)
+		links, _ := h.dht.GetLinks(hash, "4stars", StatusLive)
 		So(fmt.Sprintf("%v", links), ShouldEqual, fmt.Sprintf("[{QmWbbUf6G38hT27kmrQ5UYFbXUPTGokKvDiaQbczFYNjuN    %s}]", h.nodeIDStr))
-		links, _ = h.dht.getLinks(hash, "4stars", StatusDeleted)
+		links, _ = h.dht.GetLinks(hash, "4stars", StatusDeleted)
 		So(fmt.Sprintf("%v", links), ShouldEqual, fmt.Sprintf("[{QmYeinX5vhuA91D3v24YbgyLofw9QAxY6PoATrBHnRwbtt    %s}]", h.nodeIDStr))
 	})
 
@@ -938,7 +938,7 @@ func TestJSDHT(t *testing.T) {
 		So(header.Change.String(), ShouldEqual, profileHash.String())
 
 		// the entry should be marked as Modifed
-		data, _, _, _, err := h.dht.get(profileHash, StatusDefault, GetMaskDefault)
+		data, _, _, _, err := h.dht.Get(profileHash, StatusDefault, GetMaskDefault)
 		So(err, ShouldEqual, ErrHashModified)
 		So(string(data), ShouldEqual, profileHashStr2)
 
@@ -1031,12 +1031,12 @@ func TestJSDHT(t *testing.T) {
 
 		// the new Key should be available on the DHT
 		newKey, _ := NewHash(h.nodeIDStr)
-		data, _, _, _, err := h.dht.get(newKey, StatusDefault, GetMaskDefault)
+		data, _, _, _, err := h.dht.Get(newKey, StatusDefault, GetMaskDefault)
 		So(err, ShouldBeNil)
 		So(string(data), ShouldEqual, newPubKey)
 
 		// the old key should be marked as Modifed and we should get the new hash as the data
-		data, _, _, _, err = h.dht.get(oldKey, StatusDefault, GetMaskDefault)
+		data, _, _, _, err = h.dht.Get(oldKey, StatusDefault, GetMaskDefault)
 		So(err, ShouldEqual, ErrHashModified)
 		So(string(data), ShouldEqual, h.nodeIDStr)
